+++ test/benchmark/power.sol
@@ -11,7 +11,7 @@ contract POWER {
 	}
 	function power64(uint64 x, uint64 y) public pure returns(uint64) {
 		uint64 rtn = uint64(1);
-		for(uint64 i = uint64(0); i < y; i += 1) {
+		for(uint64 i = uint64(0); i < y; i += uint64(1)) {
 			rtn = rtn * x;
 		}
 		return rtn;
